GH-272 Fixes AnnotationSummaryBox size update

+++ viewer/viewer-awt/src/main/java/org/icepdf/ri/common/views/annotations/summary/AnnotationSummaryBox.java
@@ -344,7 +344,7 @@ public class AnnotationSummaryBox extends PopupAnnotationComponent implements Fo
     }
 
     @Override
-    public void setFontSize(final int size) {
+    public void setFontSize(final float size) {
         titleLabel.setFont(titleLabel.getFont().deriveFont(size));
         creationLabel.setFont(titleLabel.getFont().deriveFont(size));
         textArea.setFont(titleLabel.getFont().deriveFont(size));
